Hecho imágenes y font en PC
Dudas moviles:
- De donde sacamos archivos
- Flicker de la ventana al reescalar
- Barra de la ventana de la aplicacion
- Creacion de imagenes desde el juego, en start o hacerlas todas en el render
	- Pone que desde la logica, hay que hacer un resource manager o algo asi?
༼ つ ◕_◕ ༽つ

+++ 0hn0depalo/Engine/src/main/java/es/ucm/gdv/engine/Font.java
@@ -4,6 +4,6 @@ public interface Font {
     String getFilename();
 
     void setBold(boolean bold);
-    void setSize(int size);
+    void setSize(float size);
 
 }
